+++ src/degiro_portfolio/__init__.py
@@ -1,2 +1,2 @@
 """Stock price visualization application."""
-__version__ = "0.4.1"
+__version__ = "0.4.2"